+++ Changes.md
@@ -1,6 +1,11 @@
 1.5.x.x (relative to 1.5.13.0)
 =======
 
+Features
+--------
+
+- ShuffleOptions : Added a new node for shuffling options in the scene globals (#6412).
+
 Improvements
 ------------